backend/requirements.txt
python-dotenv
fastapi
openai
uvicorn
cerebras_cloud_sdk
supabase